foodlist mobile app
A user
(after obtaining height, weight and gender information) ideal weight and daily calorie intake
Calculates the amount of calories in a serving or daily total of various foods & beverages
It allows the user to create a daily meal menu by providing information about the amount and content.
A mobile application that provides.